[FG-7415] Code review
