--- flowkitAdditions/Classes/StepOutput/Double+StepOutput.swift
@@ -1,10 +1,0 @@
-//
-//  Double+StepOutput.swift
-//  flowkit-ios
-//
-//  Created by Giulio Lombardo  on 29/04/22.
-//
-
-import N26FlowKitCore
-
-extension Double: StepOutput { }
