--- flowkitAdditions/Classes/StepOutput/Float+StepOutput.swift
@@ -1,10 +1,0 @@
-//
-//  Float+StepOutput.swift
-//  flowkitAdditions
-//
-//  Created by Giulio Lombardo  on 29/04/22.
-//
-
-import N26FlowKitCore
-
-extension Float: StepOutput { }
