--- flowkitAdditions/Classes/StepOutput/Int+StepOutput.swift
@@ -1,10 +1,0 @@
-//
-//  Int+StepOutput.swift
-//  flowkitAdditions
-//
-//  Created by Giulio Lombardo  on 29/04/22.
-//
-
-import N26FlowKitCore
-
-extension Int: StepOutput { }
