+++ flowkitAdditions/Classes/StepOutput/StepOutput+Extensions.swift
@@ -1,0 +1,20 @@
+//
+//  StepOutput+Extension.swift
+//  flowkitAdditions
+//
+//  Created by Giulio Lombardo  on 29/04/22.
+//
+
+import N26FlowKitCore
+
+extension DateInterval: StepOutput { }
+
+extension Decimal: StepOutput { }
+
+extension Double: StepOutput { }
+
+extension Float: StepOutput { }
+
+extension Int: StepOutput { }
+
+extension Set: StepOutput { }
